requirements.txt
fastapi
requests
orjson
pywebio
pywebio_battery
aiofiles
websockets
pandas
click
sqlmodel
supabase
python-dotenv
fastapi-crudrouter
fastapi-cloudauth